# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "APWorld Package Index",
    "description": "Repository server for APWorld package hosting and discovery",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/packages/{name}/upload": {
      "post": {
        "tags": [
          "upload"
        ],
        "summary": "Upload Package",
        "description": "Upload a new package version.\n\nRequires authentication with 'upload' scope.\nValidates manifest and checksums.\nFirst uploader becomes package owner.",
        "operationId": "upload_package_v1_packages__name__upload_post",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          },
          {
            "name": "sha256",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sha256"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_package_v1_packages__name__upload_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/packages/{name}/{version}/yank": {
      "delete": {
        "tags": [
          "upload"
        ],
        "summary": "Yank Version",
        "description": "Yank (soft-delete) a version.\n\nYanked versions are still downloadable but marked as deprecated.\nRequires authentication and package ownership.",
        "operationId": "yank_version_v1_packages__name___version__yank_delete",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          },
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Version"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/YankRequest",
                "default": {
                  "reason": ""
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/packages/{name}/collaborators": {
      "post": {
        "tags": [
          "upload"
        ],
        "summary": "Add Collaborator",
        "description": "Add a collaborator to a package.\n\nOnly package owners can add collaborators.",
        "operationId": "add_collaborator_v1_packages__name__collaborators_post",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CollaboratorRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "upload"
        ],
        "summary": "List Collaborators",
        "description": "List all collaborators for a package.",
        "operationId": "list_collaborators_v1_packages__name__collaborators_get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/packages/{name}/collaborators/{collaborator_id}": {
      "delete": {
        "tags": [
          "upload"
        ],
        "summary": "Remove Collaborator",
        "description": "Remove a collaborator from a package.\n\nOnly package owners can remove collaborators.\nCannot remove the last owner.",
        "operationId": "remove_collaborator_v1_packages__name__collaborators__collaborator_id__delete",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          },
          {
            "name": "collaborator_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Collaborator Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_package_v1_packages__name__upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_package_v1_packages__name__upload_post"
      },
      "CollaboratorRequest": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "publisher_type": {
            "type": "string",
            "title": "Publisher Type",
            "default": "user"
          },
          "github_repository": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Github Repository"
          },
          "github_workflow": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Github Workflow"
          }
        },
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "CollaboratorRequest",
        "description": "Request body for adding/removing collaborators."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UploadResponse": {
        "properties": {
          "package": {
            "type": "string",
            "title": "Package"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "sha256": {
            "type": "string",
            "title": "Sha256"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "package",
          "version",
          "filename",
          "sha256",
          "message"
        ],
        "title": "UploadResponse",
        "description": "Response for successful upload."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "YankRequest": {
        "properties": {
          "reason": {
            "type": "string",
            "title": "Reason",
            "default": ""
          }
        },
        "type": "object",
        "title": "YankRequest",
        "description": "Request body for yanking a version."
      }
    }
  }
}